major refactoring of the questions component
